# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a data_govern use golang",
        "title": "Potato Api",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/businesses": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "业务系统列表",
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"data\" : {}, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "新增业务系统",
                "parameters": [
                    {
                        "type": "string",
                        "description": "业务系统 名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "业务系统 描述",
                        "name": "desc",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "业务系统 云端id",
                        "name": "c_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/businesses/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "删除业务系统",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "业务系统 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "businesses"
                ],
                "summary": "修改业务系统",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "业务系统 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "业务系统 名称",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "业务系统 描述",
                        "name": "desc",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "业务系统 云端id",
                        "name": "c_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/meta_databases": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta_databases"
                ],
                "summary": "数据源列表",
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"basic\" : {}, \"msg\": \"ok\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/meta_databases/{id}/meta_tables": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta_tables"
                ],
                "summary": "元数据列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "数据源 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"basic\" : {}, \"msg\": \"ok\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/vendors": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "系统厂商列表",
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"data\" : {}, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "新增系统厂商",
                "parameters": [
                    {
                        "type": "string",
                        "description": "系统厂商 名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "系统厂商 云端id",
                        "name": "c_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/vendors/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "删除系统厂商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "系统厂商 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vendors"
                ],
                "summary": "修改系统厂商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "系统厂商 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "系统厂商 名称",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "系统厂商 云端id",
                        "name": "c_id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\" : 200, \"msg\" : \"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}